+++ src/lib/utils/MdnsUtils.ts
@@ -1,13 +1,14 @@
-module.exports = function() {
-    const net = require('net')
-    const os = require('os')
-    let availableInterfaces
+import * as net from 'net'
+import * as os from 'os'
+
+const MdnsUtils = () => {
+    let availableInterfaces: string []
     
-    const checkIp = function(value) {
+    const checkIp = function(value: string) {
         return value.length > 0 && net.isIP(value)
     }
     
-    const checkInterface = function(value) {
+    const checkInterface = function(value: string) {
         if (value.length < 1) {
             return false
         }
@@ -27,44 +28,46 @@ module.exports = function() {
         return availableInterfaces.indexOf(value) > -1
     }
     
-    const checkMulticast = function(value) {
+    const checkMulticast = function(value: any) {
         return checkBoolean(value)
     }
     
-    const checkPort = function(value) {
+    const checkPort = function(value: string) {
         return value.length > 0 && checkNumber(value)
     }
     
-    const checkLoopback = function(value) {
+    const checkLoopback = function(value: any) {
         return checkBoolean(value)
     }
     
-    const checkReuseAddr = function(value) {
+    const checkReuseAddr = function(value: any) {
         return checkBoolean(value)
     }
     
-    const checkTtl = function(value) {
+    const checkTtl = function(value: string) {
         if (value.length > 0 && checkNumber(value)) {
             const ttlInt = parseInt(value)
             return ttlInt >= 0 && ttlInt <= 255
         } else return false
     }
     
-    const checkBoolean = function(value) {
+    const checkBoolean = function(value: any) {
         return typeof value === 'boolean'
     }
     
-    const checkNumber = function(value) {
+    const checkNumber = function(value: any) {
         return !isNaN(value)
     }
-    
+
     return {
-        checkInterface: checkInterface,
-        checkIp: checkIp,
-        checkMulticast: checkMulticast,
-        checkPort: checkPort,
-        checkLoopback: checkLoopback,
-        checkReuseAddr: checkReuseAddr,
-        checkTtl: checkTtl,
+        checkInterface,
+        checkIp,
+        checkMulticast,
+        checkPort,
+        checkLoopback,
+        checkReuseAddr,
+        checkTtl,
     }
 }
+
+module.exports = MdnsUtils