+++ src/lib/utils/index.js
@@ -1,17 +1,15 @@
 module.exports = function(node) {
-    const ServiceUtils = require('./ServiceUtils.js')(node)
-    const BridgeUtils = require('./BridgeUtils.js')(node)
-    const AccessoryUtils = require('./AccessoryUtils.js')(node)
-    const CharacteristicUtils = require('./CharacteristicUtils.js')(node)
-    const MdnsUtils = require('./MdnsUtils.js')()
-    const EnvironmentUtils = require('./EnvironmentUtils')()
+    const ServiceUtils = require('./ServiceUtils')(node)
+    const BridgeUtils = require('./BridgeUtils')(node)
+    const AccessoryUtils = require('./AccessoryUtils')(node)
+    const CharacteristicUtils = require('./CharacteristicUtils')(node)
+    const MdnsUtils = require('./MdnsUtils')()
     
     return {
         ServiceUtils: ServiceUtils,
         BridgeUtils: BridgeUtils,
         AccessoryUtils: AccessoryUtils,
         CharacteristicUtils: CharacteristicUtils,
-        MdnsUtils: MdnsUtils,
-        EnvironmentUtils: EnvironmentUtils,
+        MdnsUtils: MdnsUtils
     }
 }